projects:ad405x_iio: Refactoring of the project
The refactoring of ad405x has been done to improve the readability and
usage of the code. As AD405x is an umbrella term for multiple generics,
as support increases the code tends to be cluttered.
