+++ projects/ad405x_iio/STM32/sdp_k1/.extSettings
@@ -2,7 +2,7 @@
 HeaderPath=../../app;../../../../libraries/no-OS/util;../../../../libraries/no-OS/include;../../../../libraries/no-OS/drivers/platform/stm32;../../../../libraries/no-OS/iio;../../../../libraries/no-OS/drivers/api;../../../../libraries/precision-converters-library/board_info/;../../../../libraries/no-OS/drivers/eeprom/24xx32a/;../../../../libraries/precision-converters-library/sdp_k1_sdram/;../../../../libraries/precision-converters-library/common/;../../../../libraries/no-OS/drivers/adc/ad405x/;
 
 [Groups]
-app/=../../app/main.c;../../app/ad405x_iio.c;../../app/ad405x_iio.h;../../app/app_config.h;../../app/app_config.c;../../app/app_config_stm32.c;../../app/app_config_stm32.h;../../app/stm32_gpio_irq_generated.c;../../app/ad405x_user_config.c;../../app/ad405x_user_config.h;
+app/=../../app/main.c;../../app/ad405x_iio.c;../../app/ad405x_iio.h;../../app/app_config.h;../../app/app_config.c;../../app/app_config_stm32.c;../../app/app_config_stm32.h;../../app/stm32_gpio_irq_generated.c;../../app/ad405x_user_config.c;../../app/ad405x_user_config.h;../../app/app_support.h;../../app/app_support.c;../../app/ad405x_support.c;
 
 app/libraries/precision-converters-library/board_info/=../../../../libraries/precision-converters-library/board_info/board_info.c;../../../../libraries/precision-converters-library/board_info/board_info.h;
